refactored code to make a separate project for concordium protocol types
