+++ contracts/concordium-protocols/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "concordium-protocols"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+concordium-std = { version = "8.1.0", default-features = false }
+concordium-cis2 = { version = "5.1.0", default-features = false }
