+++ contracts/concordium-protocols/src/concordium_cis3.rs
