+++ contracts/concordium-protocols/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod concordium_cis2_security;
+pub mod concordium_cis3;
+pub mod concordium_cis2_ext;
