+++ contracts/integration-tests/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 concordium-std = { version = "8.1.0", default-features = false }
 concordium-cis2 = { version = "5.1.0", default-features = false }
+concordium-protocols = { path = "../concordium-protocols" }
 concordium-rwa-identity-registry = { path = "../identity-registry" }
 concordium-rwa-compliance = { path = "../compliance" }
 concordium-rwa-security-nft = { path = "../security-nft" }
