+++ contracts/utils/Cargo.toml
@@ -14,6 +14,7 @@ wee_alloc = ["concordium-std/wee_alloc"]
 concordium-std = { version = "8.1.0", default-features = false }
 concordium-cis2 = { version = "5.1.0", default-features = false }
 hex = { version = "0.4.3" }
+concordium-protocols = { path = "../concordium-protocols" }
 
 [lib]
 crate-type = ["cdylib", "rlib"]
