+++ contracts/utils/src/lib.rs
@@ -7,7 +7,6 @@ pub mod cis2_types;
 pub mod clients;
 pub mod common_types;
 pub mod compliance_types;
-pub mod concordium_cis3;
 pub mod holders_security_state;
 pub mod holders_state;
 pub mod sponsor_types;
@@ -15,4 +14,6 @@ pub mod sponsors_state;
 pub mod token_deposits_state;
 pub mod tokens_security_state;
 pub mod tokens_state;
-pub mod concordium_cis2_security;
+pub use concordium_protocols::concordium_cis2_ext;
+pub use concordium_protocols::concordium_cis3;
+pub use concordium_protocols::concordium_cis2_security;
