+++ contracts/utils/src/tokens_state.rs
@@ -1,68 +1,10 @@
 use concordium_cis2::*;
-use concordium_std::{ops, *};
+use concordium_std::*;
 
 #[derive(Serial, DeserialWithState)]
 #[concordium(state_parameter = "S")]
 pub struct TokensState<T, TTokenState, S> {
     tokens: StateMap<T, TTokenState, S>,
-}
-
-/// Trait representing a token amount.
-///
-/// This trait is used to define the behavior of token amounts.
-pub trait IsTokenAmount:
-    concordium_cis2::IsTokenAmount
-    + PartialOrd
-    + ops::SubAssign
-    + Copy
-    + ops::AddAssign
-    + ops::Sub<Output = Self> {
-    /// Returns the zero value of the token amount.
-    fn zero() -> Self;
-
-    /// Returns the maximum value of the token amount.
-    /// This should return `1` for NFTs.
-    fn max_value() -> Self;
-
-    /// Subtracts the given amount from self. Returns None if the amount is too
-    /// large.
-    ///
-    /// # Arguments
-    ///
-    /// * `other` - The amount to subtract.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Some(())` if the subtraction was successful, `None` otherwise.
-    fn checked_sub_assign(&mut self, other: Self) -> Option<()> {
-        if other.le(self) {
-            self.sub_assign(other);
-            Some(())
-        } else {
-            None
-        }
-    }
-
-    /// Adds the given amount to self. Returns None if the amount is too large.
-    ///
-    /// # Arguments
-    ///
-    /// * `other` - The amount to add.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Some(())` if the addition was successful, `None` otherwise.
-    fn checked_add_assign(&mut self, other: Self) -> Option<()> {
-        if other.le(&Self::max_value().sub(*self)) {
-            self.add_assign(other);
-            Some(())
-        } else {
-            None
-        }
-    }
-
-    /// Returns true if the amount is zero.
-    fn is_zero(&self) -> bool { self.eq(&Self::zero()) }
 }
 
 pub enum TokenStateError {
